@@ -2,6 +2,7 @@ package dagger
 
 import (
 	"context"
+	"fmt"
 	"path"
 	"path/filepath"
 
@@ -11,24 +12,31 @@ import (
 	"github.com/rs/zerolog/log"
 
 	"dagger.cloud/go/dagger/cc"
+	"dagger.cloud/go/stdlib"
 )
 
 // Build a cue configuration tree from the files in fs.
 func CueBuild(ctx context.Context, fs FS, args ...string) (*cc.Value, error) {
-	lg := log.Ctx(ctx)
-
-	// The CUE overlay needs to be prefixed by a non-conflicting path with the
-	// local filesystem, otherwise Cue will merge the Overlay with whatever Cue
-	// files it finds locally.
-	const overlayPrefix = "/config"
+	var (
+		err error
+		lg  = log.Ctx(ctx)
+	)
 
 	buildConfig := &cueload.Config{
-		Dir:     overlayPrefix,
-		Overlay: map[string]cueload.Source{},
+		// The CUE overlay needs to be prefixed by a non-conflicting path with the
+		// local filesystem, otherwise Cue will merge the Overlay with whatever Cue
+		// files it finds locally.
+		Dir: "/config",
 	}
-	buildArgs := args
 
-	err := fs.Walk(ctx, func(p string, f Stat) error {
+	// Start by creating an overlay with the stdlib
+	buildConfig.Overlay, err = stdlib.Overlay(buildConfig.Dir)
+	if err != nil {
+		return nil, err
+	}
+
+	// Add the config files on top of the overlay
+	err = fs.Walk(ctx, func(p string, f Stat) error {
 		lg.Debug().Str("path", p).Msg("Compiler.Build: processing")
 		if f.IsDir() {
 			return nil
@@ -38,16 +46,16 @@ func CueBuild(ctx context.Context, fs FS, args ...string) (*cc.Value, error) {
 		}
 		contents, err := fs.ReadFile(ctx, p)
 		if err != nil {
-			return errors.Wrap(err, p)
+			return fmt.Errorf("%s: %w", p, err)
 		}
-		overlayPath := path.Join(overlayPrefix, p)
+		overlayPath := path.Join(buildConfig.Dir, p)
 		buildConfig.Overlay[overlayPath] = cueload.FromBytes(contents)
 		return nil
 	})
 	if err != nil {
 		return nil, err
 	}
-	instances := cueload.Instances(buildArgs, buildConfig)
+	instances := cueload.Instances(args, buildConfig)
 	if len(instances) != 1 {
 		return nil, errors.New("only one package is supported at a time")
 	}
