@@ -2,7 +2,6 @@ package dagger
 
 import (
 	"context"
-	"os"
 
 	"cuelang.org/go/cue"
 	cueflow "cuelang.org/go/tools/flow"
@@ -91,36 +90,9 @@ func (env *Env) SetInput(i *cc.Value) error {
 	)
 }
 
-func stdlibLoader() (*cc.Value, error) {
-	if dev := os.Getenv("DAGGER_DEV_STDLIB"); dev != "" {
-		v, err := cc.Compile("stdlib.cue", `
-			do: "local"
-			dir: string
-			include: ["cue.mod/pkg"]
-		`)
-		if err != nil {
-			return nil, err
-		}
-		return v.MergeTarget(dev, "dir")
-	}
-	return cc.Compile("stdlib.cue", `
-		do: "fetch-git"
-		remote: "https://github.com/blocklayerhq/dagger-stdlib"
-		ref: "0625677b5aec1162621ad18fbd7b90dc9d7d54e5"
-	`)
-}
-
 // Update the base configuration
 func (env *Env) Update(ctx context.Context, s Solver) error {
 	p := NewPipeline(s, nil)
-	// always inject stdlib in cue.mod/pkg
-	stdlib, err := stdlibLoader()
-	if err != nil {
-		return err
-	}
-	if err := p.Do(ctx, stdlib); err != nil {
-		return err
-	}
 	// execute updater script
 	if err := p.Do(ctx, env.updater); err != nil {
 		return err
@@ -189,10 +161,6 @@ func (env *Env) LocalDirs() map[string]string {
 	)
 	// 2. Scan the environment updater
 	localdirs(env.Updater())
-	// 3. In dev mode, always include dev stdlib directory
-	if dev := os.Getenv("DAGGER_DEV_STDLIB"); dev != "" {
-		dirs[dev] = dev
-	}
 	return dirs
 }
 
@@ -337,14 +305,7 @@ func (env *Env) Compute(ctx context.Context, s Solver) error {
 			}
 			v := cc.Wrap(t.Value(), flowInst)
 			p := NewPipeline(s, NewFillable(t))
-			err := p.Do(ctx, v)
-			if err == ErrAbortExecution {
-				// Pipeline was partially executed
-				// FIXME: tell user which inputs are missing (by inspecting references)
-				lg.Warn().Msg("pipeline was partially executed because of missing inputs")
-				return nil
-			}
-			return err
+			return p.Do(ctx, v)
 		}), nil
 	}
 	// Orchestrate execution with cueflow
